Hóa đơn 99%, Trang Chủ 99%

+++ ProjectLuanVan/TrainingCenters/InterfacesApi/IPhieuThuChi.cs
@@ -5,7 +5,7 @@ namespace TrainingCenters.InterfacesApi
     {
         Task<object> LoadingDataTableView(PhieuThuChi item, int skip, int take);
         Task<ICollection<PhieuThuChi>> Search(PhieuThuChi item);
-        Task<int> SearchTongTien(PhieuThuChi item);
+        Task<double> SearchTongTien(PhieuThuChi item);
         Task<int> SearchCount(PhieuThuChi item);
         Task<ICollection<PhieuThuChi>> GetAll();
         Task<PhieuThuChi> GetById(int id);
