+++ WebApi_ManagementCenters/Data/InterfacesData/IPhieuThuChi.cs
@@ -11,7 +11,7 @@ namespace Data.InterfacesData
     {
         object LoadingDataTableView(PhieuThuChi item, int skip, int take);
         Task<ICollection<PhieuThuChi>> Search(PhieuThuChi item);
-        Task<int> SearchTongTien(PhieuThuChi item);
+        Task<double> SearchTongTien(PhieuThuChi item);
         Task<int> SearchCount(PhieuThuChi item);
         Task<ICollection<PhieuThuChi>> GetAll();
         Task<PhieuThuChi> GetById(int id);
